requirements.txt
frida
python-dotenv
requests
watchdog